fix delete size and color

+++ src/main/java/yourstyle/com/shope/controller/admin/ColorController.java
@@ -1,15 +1,20 @@
 package yourstyle.com.shope.controller.admin;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.ui.ModelMap;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import jakarta.validation.Valid;
 import yourstyle.com.shope.model.Color;
 import yourstyle.com.shope.service.ColorService;
+import yourstyle.com.shope.service.impl.ColorServiceImpl;
 
 @Controller
 @RequestMapping("/admin/colors")
@@ -17,6 +22,8 @@ public class ColorController {
 
     @Autowired
     private ColorService colorService;
+    @Autowired
+    ColorServiceImpl colorServiceImpl;
 
     @PostMapping("/save")
     public String saveColor(@Valid @ModelAttribute("color") Color color, BindingResult result, Model model) {
@@ -28,12 +35,13 @@ public class ColorController {
         return "redirect:/admin/products/add";
     }
 
-    @GetMapping("/delete/{id}")
-    public String deleteColor(@PathVariable("id") Integer id, ModelMap model) {
+    @DeleteMapping("/delete/{id}")
+    public ResponseEntity<String> deleteColor(@PathVariable("id") Integer id) {
+        if (colorServiceImpl.isColorInUse(id)) {
+            return ResponseEntity.status(HttpStatus.CONFLICT).body("Không thể xóa! Màu đang được sử dụng!");
+        }
         colorService.deleteById(id);
-        model.addAttribute("messageType", "success");
-        model.addAttribute("messageContent", "Xóa thành công");
-        return "redirect:/admin/products/add"; // Chuyển hướng về trang danh sách
+        return ResponseEntity.ok("Xóa màu thành công!");
     }
-    
+
 }
